Value rework
Intermediate representation of the values (as enum `Value`) is not so useful for bindings but it brings unnecessary overhead in the form of twice conversion (`JSValue` -> `Value` -> `RustType` and back).
I turned the `Value` to be a wrapper around `JSValue` so the values converts directly.
The `JsRef` type is removed because the `Value` yet behaves in same manner. Other referenced entities (`String`, `Symbol`, `Object`, `Array`, `Function`) itself wraps the `Value` type and uses transparent representation so we may cast references between each of that types and the `Value`.
Also I taken out `handle_exception` from `Value::from_js_value*()` to avoid double checks for exceptions and also unnecessary unwrapings.